nim.vim; properly get salve.vim to not auto-start a REPL
I assumed setting that setting to 0 would do the trick, but it was still
auto-starting the REPL. I had a look at the source code for salve.vim
and it turns out that just having that value set to anything (even 0)
enables that behavior. It's opt in, so not setting it at all works.