experiment: turn off classpath.vim
context: dominic munroe mentioned in clojurians #vim that he doesn't use
classpath.vim, and things tend to be faster that way as long as you
start a REPL

Now I don't remember why I ever enabled classpath.vim to begin
with. The README for classpath.vim says it sets the 'path' for JVM
languages to match the classpath so that commands like :find and gf
work as designed. But I never use :find or gf, so maybe I don't need
classpath.vim after all.

Can always revert this commit if this ends up being a mistake.