always use `abort` when defining a function in vim
When something goes wrong in the middle of a function in vim, it's almost always
the case that I don't want to proceed to do the rest of the function. I want it
to throw an error in my face and stop.

I can see how there might be some situations where I want to try to do
something, and if it fails, I want to proceed to do the rest of the things that
I'm doing in that function. In those situations, I'd rather have an explicit
try/catch so that the intention is obvious. I don't think any of the functions
in this commit where I'm adding `abort` fall under that umbrella, but if they
are, I'll find out and at that point, I can go in and add try/catch as needed.