go back to using clojure.data/diff + clojure.pprint/pprint
lambdaisland/deep-diff is really cool, but it's lacking a critical feature,
which is hiding the parts of the diff that are the same. The output is really
easy to read for small files, but difficult for large files because you have to
scan through the entire file looking for difference markers.

I dug through the deep-diff repo a bit trying to see if they support this
feature via an option, etc. and it looks like they don't yet:

https://github.com/lambdaisland/deep-diff/issues/13