bootstrap_ubuntu.sh: Update Neovim installation steps
python-pip is no longer available in the apt repos out of the box. Seems in
general that the community is moving away from supporting Python 2. I'm just
going to go without the Python 2 Neovim provider and see how far I get.

If it turns out that I find that I really do need the Python 2 Neovim provider,
then I'll figure out how to add it and update the script accordingly.