Aggressively reformat js/ts/tsx files
I think I like this. Can adjust as time goes by, if needed.

One thing that's a little weird is that it's still possible, sometimes, for the
InsertLeave and TextChanged hooks not to trigger a re-format after I've changed
code. Then, when I save the file, it triggers the re-format, but then it appears
to save the un-formatted version and leave the formatting changes unsaved...

But I still feel like this is nice, overall. Will see how I feel after using it
for a while.