Put cmp sources in the same group
The way I had it before, each source was in its own group, which I think
means that the sources are tried in order, and it was stopping as soon
as it got results from any of the sources. I was seeing that whenever
Copilot returned a result (which it usually does), I would only see the
Copilot result, and the rest would be suppressd, which is not ideal.

I like it this way better, because it will show results from all
sources, and I can choose which one I want. Sometimes, I want to see LSP
and/or buffer results and have the option to select them as well.